+++ settings.py
@@ -27,7 +27,7 @@ APP_KITS = {
             'task2_tr1',
             'task2_results',
             'task3_tr13',
-            'static_portfolio_results',
+            'task3_results_tr13',
             'task4_tr13',
             'task4_results',
             'crt',
@@ -50,7 +50,7 @@ APP_KITS = {
             'task2_tr2',
             'task2_results',
             'task3_tr2',
-            'static_portfolio_results',
+            'task3_results_tr13',
             'task4_tr2',
             'task4_results',
             'crt',
@@ -73,7 +73,7 @@ APP_KITS = {
             'task2_tr1',
             'task2_results',
             'task3_tr13',
-            'static_portfolio_results',
+            'task3_results_tr13',
             'task4_tr13',
             'task4_results',
             'crt',
@@ -109,7 +109,7 @@ SESSION_CONFIGS = [
                       'task2_results',
                       'task3_tr13',
                       'task3_tr2',
-                      'static_portfolio_results',
+                      'task3_results_tr13',
                       'task4_tr13',
                       'task4_tr2',
                       'task4_results',
