+++ task3_results_tr13/templates/task3_results_tr13/MyPage.html
@@ -1,0 +1,13 @@
+{% extends "global/Page.html" %}
+{% load otree static %}
+
+{% block title %}
+    Page title
+{% endblock %}
+
+{% block content %}
+
+    {% formfields %}
+    {% next_button %}
+
+{% endblock %}
